Certificates refactoring
