+++ portfolio/src/app/services/index.ts
@@ -2,3 +2,4 @@ export { ProjectsService } from './projects.service';
 export { ScrollPositionService } from './scroll-position.service';
 export { TranslationService } from './translation.service';
 export { CvPageService } from './cv-page.service';
+export { CertificatesService } from './certificates.service';
