+++ portfolio/src/app/services/projects.service.ts
@@ -9,28 +9,17 @@ import { Language, Util } from '../enums';
   providedIn: 'root'
 })
 export class ProjectsService {
-  // Remove the readonly projectsUrl, use a getter instead
-
   constructor(private http: HttpClient, private translationService: TranslationService) { }
 
-  /**
-   * Get the correct projects file path based on the current language
-   */
   private get projectsUrl(): string {
     const lang = this.translationService.currentLangValue || Language.EN;
     return `/projects/${lang}.json`;
   }
 
-  /**
-   * Fetch all projects from the JSON file for the current language
-   */
   public getProjects(): Observable<Project[]> {
     return this.http.get<Project[]>(this.projectsUrl);
   }
 
-  /**
-   * Fetch a specific project by ID
-   */
   public getProjectById(id: string): Observable<Project | undefined> {
     return this.getProjects().pipe(
       take(Util.DEFAULT_TAKE),
