+++ portfolio/src/app/shared/index.ts
@@ -1,6 +1,5 @@
 export { ContainerComponent } from './container/container.component';
 export { TagComponent } from './tag/tag.component';
-export { CertificatesCarouselComponent } from './certificates-carousel/certificates-carousel.component';
 export { ScrollToTopComponent } from './scroll-to-top/scroll-to-top.component';
 export { FadeInDirective } from './animations/fade-in.directive';
 export { SafeUrlPipe } from './pipes/safe-url.pipe';
